app/routes/main.py
from flask import Blueprint, redirect, url_for
from flask_security import current_user

bp = Blueprint('main', __name__)

@bp.route('/')
def index():
    if current_user.is_authenticated:
        return redirect(url_for('dashboard.index'))
    return redirect(url_for('security.login'))
