requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Security-Too==5.3.2
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
email-validator==2.1.0.post1
python-dotenv==1.0.1
bcrypt==4.0.1  # Using an older version to fix compatibility issue
passlib==1.7.4
pandas==2.2.1  # For CSV handling
bleach==6.1.0
